Coin UI is now updated both ways.
There was a weird bug where after clicking a button,
Using spacebar would result in being registered as a button click.
This fires the OnClick event again.
Going to the InputManager in project settings,
Then changing the altpositivebutton from space to enter fixed the issue.

TODO: create UI object for when the player tries to buy something, but doesn't have enough coins.
Will do this in a later feature.